+++ enrollment-server-onboarding/src/main/java/com/wultra/app/onboardingserver/controller/api/ConfigurationController.java
@@ -45,7 +45,7 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static net.logstash.logback.argument.StructuredArguments.kv;
+import static com.wultra.app.onboardingserver.common.logging.StructuredLogging.*;
 
 /**
  * Configuration controller.
@@ -73,7 +73,7 @@ public class ConfigurationController {
             @Parameter(hidden = true) final EncryptionContext encryptionContext) throws PowerAuthEncryptionException, InvalidRequestObjectException {
 
         final String processType = request.getRequestObject().processType();
-        logger.info("", kv("action", "fetchConfiguration"), kv("state", "initiated"), kv("processType", processType));
+        logger.info("", action("fetchConfiguration"), stateInitiated(), kv("processType", processType));
 
         if (encryptionContext == null) {
             throw new PowerAuthEncryptionException("ECIES decryption failed");
@@ -86,8 +86,8 @@ public class ConfigurationController {
                 .otpResendPeriodSeconds(onboardingConfig.getOtpResendPeriod().getSeconds())
                 .build();
 
-        logger.info("", kv("action", "fetchConfiguration"), kv("state", "succeeded"));
-        logger.debug("", kv("action", "fetchConfiguration"), kv("state", "succeeded"), kv("result", result));
+        logger.info("", action("fetchConfiguration"), stateSucceeded());
+        logger.debug("", action("fetchConfiguration"), stateSucceeded(), kv("result", result));
 
         return new ObjectResponse<>(result);
     }