+++ enrollment-server-onboarding/src/main/java/com/wultra/app/onboardingserver/controller/api/PrivateClientController.java
@@ -30,7 +30,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import static net.logstash.logback.argument.StructuredArguments.kv;
+import static com.wultra.app.onboardingserver.common.logging.StructuredLogging.*;
 
 /**
  * Controller to acknowledge asynchronous actions. These endpoints are private and secured.
@@ -52,10 +52,10 @@ class PrivateClientController {
      */
     @PostMapping("approve")
     public AcknowledgeApproveClientResponse acknowledgeApproveClient(final @Valid @RequestBody AcknowledgeApproveClientRequest request) {
-        logger.info("", kv("action", "acknowledgeApproveClient"), kv("state", "initiated"), kv("processId", request.processId()));
+        logger.info("", action("acknowledgeApproveClient"), stateInitiated(), kv("processId", request.processId()));
 
         if (request.approvalResult() == AcknowledgeApproveClientRequest.ApprovalResult.WAIT) {
-            logger.info("", kv("action", "acknowledgeApproveClient"), kv("state", "skipped"), kv("reason", "approvalResult is WAIT"));
+            logger.info("", action("acknowledgeApproveClient"), state("skipped"), kv("reason", "approvalResult is WAIT"));
             return AcknowledgeApproveClientResponse.builder()
                     .result(AcknowledgeApproveClientResponse.Result.OK)
                     .build();
@@ -64,9 +64,9 @@ class PrivateClientController {
         final AcknowledgeApproveClientResponse response = acknowledgeService.acknowledgeApproveClient(request);
 
         if (response.result() == AcknowledgeApproveClientResponse.Result.OK) {
-            logger.info("", kv("action", "acknowledgeApproveClient"), kv("state", "succeeded"));
+            logger.info("", action("acknowledgeApproveClient"), stateSucceeded());
         } else {
-            logger.info("", kv("action", "acknowledgeApproveClient"), kv("state", "failed"), kv("reason", response.resultReason()));
+            logger.info("", action("acknowledgeApproveClient"), stateFailed(), kv("reason", response.resultReason()));
         }
 
         return response;
@@ -74,10 +74,10 @@ class PrivateClientController {
 
     @PostMapping("evaluate")
     public AcknowledgeEvaluationClientResponse acknowledgeEvaluationClient(final @Valid @RequestBody AcknowledgeEvaluationClientRequest request) {
-        logger.info("", kv("action", "acknowledgeEvaluationClient"), kv("state", "initiated"), kv("processId", request.processId()));
+        logger.info("", action("acknowledgeEvaluationClient"), stateInitiated(), kv("processId", request.processId()));
 
         if (request.evaluationResult() == AcknowledgeEvaluationClientRequest.EvaluationResult.WAIT) {
-            logger.info("", kv("action", "acknowledgeEvaluationClient"), kv("state", "skipped"), kv("reason", "evaluationResult is WAIT"));
+            logger.info("", action("acknowledgeEvaluationClient"), state("skipped"), kv("reason", "evaluationResult is WAIT"));
             return AcknowledgeEvaluationClientResponse.builder()
                     .result(AcknowledgeEvaluationClientResponse.Result.OK)
                     .build();
@@ -86,9 +86,9 @@ class PrivateClientController {
         final AcknowledgeEvaluationClientResponse response = acknowledgeService.acknowledgeEvaluationClient(request);
 
         if (response.result() == AcknowledgeEvaluationClientResponse.Result.OK) {
-            logger.info("", kv("action", "acknowledgeEvaluationClient"), kv("state", "succeeded"));
+            logger.info("", action("acknowledgeEvaluationClient"), stateSucceeded());
         } else {
-            logger.info("", kv("action", "acknowledgeEvaluationClient"), kv("state", "failed"), kv("reason", response.resultReason()));
+            logger.info("", action("acknowledgeEvaluationClient"), stateFailed(), kv("reason", response.resultReason()));
         }
 
         return response;