+++ enrollment-server-onboarding/src/main/java/com/wultra/app/onboardingserver/impl/service/userdatastore/NoOpUserDataStoreService.java
@@ -22,7 +22,7 @@ import lombok.extern.slf4j.Slf4j;
 
 import java.util.List;
 
-import static net.logstash.logback.argument.StructuredArguments.kv;
+import static com.wultra.app.onboardingserver.common.logging.StructuredLogging.*;
 
 /**
  * Empty implementation {@link UserDataStoreService}.
@@ -34,6 +34,6 @@ class NoOpUserDataStoreService implements UserDataStoreService {
 
     @Override
     public void storeDocumentData(final List<DocumentCreateRequest> documentRequests) {
-        logger.info("", kv("action", "storeDocumentData"), kv("state", "skipped"));
+        logger.info("", action("storeDocumentData"), state("skipped"));
     }
 }